Change tests to expected behavior
